Added `extra` Project Folder
Changes Made
• Added `/extra` directory for project management
• Added `THE-GIST.md` to `/extra`

Progress Made
• Checked compiler compilation (Successful)
• Established stage0 and stage1 toolchain
• Began forming project overview and outlook

Problems Found:
• Unable to establish stage2 toolchain (User error)
• Possible development delays from compilation time (Hardware issue, User error)
• Potential violation of basic developer practices (User error, Skill issue)

___

`/extra` will be used to manage and record this project by storing change logs and temporary documentation. In the unlikely event that a pull request is made, `/extra` and all of its content should be excluded from the request.

`THE-GIST.md` is intended to serve as a project overview. It should outline the basic ideas and conditions of this project. Hopefully, it will change very little in content over the course of this project. Semantic changes will likely occur, but those are trivial and are expected from the beginning.

I have confirmed that the compiler is properly running on my machine and will begin working on the `class` keyword once I have completed the project outline. Performing checks and tests has been particularly slow, but it provide ample time to write reports and logs like this, so there's some benefit to that.